Moves images & renames for consistency
This PR moves the images folders for these 2 metrics models to be consistent with other metrics models:
 - Project Engagement
 - Community Activity

This PR also fixes broken images links for these metrics models and renames the Project Engagement model to be more consistent with the naming of other metrics models.

Note to website team that we will need to update the path to these metrics on the site.